Store activations in full precision and update examples
Drop .half() casts on activations/logits/loss so outputs stay in
their original dtype, guard tokens against None, and remove the
unused vision examples utility.

+++ src/nnact/_model/_hooked.py
@@ -169,9 +169,7 @@ class HookedModel(nn.Module):
                                 f"is not a Tensor, got {type(output)}"
                             )
 
-                        self._activations[layer_key] = (
-                            tensor.half().detach().cpu().numpy()
-                        )
+                        self._activations[layer_key] = tensor.detach().cpu().numpy()
 
                     return hook
